+++ lib/active_record/cursor_paginator.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'active_record'
+require 'set'
 require_relative 'cursor_paginator/version'
 
 module ActiveRecord
@@ -168,7 +169,7 @@ module ActiveRecord
       def cursor_for_record(record)
         unencoded_cursor = @fields.map do |field|
           field_name = field.keys.first
-          value = if record.class.defined_enums.key?(field_name.to_s)
+          value = if record.class.defined_enums.has_key?(field_name.to_s)
                     # For enum columns, get the raw integer value from the database
                     record.read_attribute_before_type_cast(field_name)
                   else
@@ -268,14 +269,43 @@ module ActiveRecord
       def build_filter_query(sorted_relation, op, current_field, prev_fields)
         relation = sorted_relation
         prev_fields.each do |col, val|
+          col_key = col
           col = @aliases[col] if @aliases.has_key? col
           col = qualify_field_if_needed(col)
-          relation = relation.where("#{col} = ?", val)
+          relation = relation.where("#{col} = ?", cast_cursor_value(col_key, val))
         end
         col, val = current_field
+        col_key = col
         col = @aliases[col] if @aliases.has_key? col
         col = qualify_field_if_needed(col)
-        relation.where("#{col} #{op} ?", val)
+        relation.where("#{col} #{op} ?", cast_cursor_value(col_key, val))
+      end
+
+      def cast_cursor_value(col_name, val)
+        return val unless val.is_a?(String)
+
+        resolved_expr = @aliases[col_name.to_s]
+        resolved_col = (resolved_expr || col_name.to_s).split('.').last
+        type = @relation.klass.type_for_attribute(resolved_col)
+
+        return val unless %i[datetime time timestamp].include?(type.type)
+
+        warn_join_column_cast(col_name, resolved_expr) if resolved_expr&.match?(/\A\w+\.\w+\z/)
+
+        type.cast(val)
+      end
+
+      def warn_join_column_cast(alias_name, expression)
+        @_warned_join_casts ||= Set.new
+        return if @_warned_join_casts.include?(alias_name)
+
+        @_warned_join_casts << alias_name
+
+        msg = "[CursorPaginator] Cursor column '#{alias_name}' resolves to '#{expression}' " \
+              "(join column). Type is inferred from #{@relation.klass.name} and may be " \
+              'inaccurate. Timezone-aware datetime cursors on joined tables may not work correctly.'
+        logger = ActiveRecord::Base.logger
+        logger ? logger.warn(msg) : Kernel.warn(msg)
       end
 
       # parse aliases from select values